testes de unidade

+++ src/main/java/com/rodrigo/helpdesk/repositories/ChamadoRepository.java
@@ -2,8 +2,8 @@ package com.rodrigo.helpdesk.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.rodrigo.helpdesk.domain.Chamado;
+import com.rodrigo.helpdesk.model.Chamado;
 
-public interface ChamadoRepository extends JpaRepository<Chamado, Integer> {
+public interface ChamadoRepository extends JpaRepository<Chamado, Long> {
 	
 }
