+++ src/main/java/com/rodrigo/helpdesk/repositories/ClienteRepository.java
@@ -2,8 +2,8 @@ package com.rodrigo.helpdesk.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.rodrigo.helpdesk.domain.Cliente;
+import com.rodrigo.helpdesk.model.Cliente;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
+public interface ClienteRepository extends JpaRepository<Cliente, Long> {
 	
 }
