+++ src/main/java/com/rodrigo/helpdesk/repositories/PessoaRepository.java
@@ -4,9 +4,9 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.rodrigo.helpdesk.domain.Pessoa;
+import com.rodrigo.helpdesk.model.Pessoa;
 
-public interface PessoaRepository extends JpaRepository<Pessoa, Integer> {
+public interface PessoaRepository extends JpaRepository<Pessoa, Long> {
 	Optional<Pessoa> findByCpf(String cpf);
 	Optional<Pessoa> findByEmail(String email);
 }
