+++ src/main/java/com/rodrigo/helpdesk/repositories/TecnicoRepository.java
@@ -2,8 +2,8 @@ package com.rodrigo.helpdesk.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.rodrigo.helpdesk.domain.Tecnico;
+import com.rodrigo.helpdesk.model.Tecnico;
 
-public interface TecnicoRepository extends JpaRepository<Tecnico, Integer> {
+public interface TecnicoRepository extends JpaRepository<Tecnico, Long> {
 	
 }
